fixed tests task
