+++ tasks/__init__.py
@@ -1,0 +1,5 @@
+"""
+Invoke entrypoint, import here all the tasks we want to make available
+"""
+from .go import fmt, lint, vet
+from .test import test
